mod: tickをfloat64にしてcustomEffectのexampleを追加

+++ sceneManager/game/circularEffect.go
@@ -34,7 +34,7 @@ func (cfce *circularFocusClosingEffect) Draw(sceneEffect SceneEffect, w, h int, 
 	x := math.Max(float64(sceneEffect.Focus.X), float64(w-sceneEffect.Focus.X))
 	y := math.Max(float64(sceneEffect.Focus.Y), float64(h-sceneEffect.Focus.Y))
 	r := math.Sqrt(math.Pow(x, 2) + math.Pow(y, 2))
-	d := (r - r*(float64(sceneEffect.Tick)/float64(sceneEffect.Frame))) * 2
+	d := (r - r*(sceneEffect.Tick/sceneEffect.Frame)) * 2
 
 	circleImage := generateCircleImage(sceneEffect, w, h, d)
 	screen.DrawImage(circleImage, nil)
@@ -46,7 +46,7 @@ func (cfoe *circularFocusOpeningEffect) Draw(sceneEffect SceneEffect, w, h int, 
 	x := math.Max(float64(sceneEffect.Focus.X), float64(w-sceneEffect.Focus.X))
 	y := math.Max(float64(sceneEffect.Focus.Y), float64(h-sceneEffect.Focus.Y))
 	r := math.Sqrt(math.Pow(x, 2) + math.Pow(y, 2))
-	d := r * (float64(sceneEffect.Tick) / float64(sceneEffect.Frame)) * 2
+	d := r * (sceneEffect.Tick / sceneEffect.Frame) * 2
 
 	circleImage := generateCircleImage(sceneEffect, w, h, d)
 	screen.DrawImage(circleImage, nil)
