+++ sceneManager/game/fadeEffect.go
@@ -7,7 +7,7 @@ import (
 type fadeInEffect struct{}
 
 func (fie *fadeInEffect) Draw(sceneEffect SceneEffect, w, h int, screen *ebiten.Image) {
-	alpha := 255 - 255*(float64(sceneEffect.Tick)/float64(sceneEffect.Frame))
+	alpha := 255 - 255*(sceneEffect.Tick/sceneEffect.Frame)
 	img := generateFadeImage(sceneEffect, w, h, alpha)
 	screen.DrawImage(img, nil)
 }
@@ -15,7 +15,7 @@ func (fie *fadeInEffect) Draw(sceneEffect SceneEffect, w, h int, screen *ebiten.
 type fadeOutEffect struct{}
 
 func (foe *fadeOutEffect) Draw(sceneEffect SceneEffect, w, h int, screen *ebiten.Image) {
-	alpha := 255 * (float64(sceneEffect.Tick) / float64(sceneEffect.Frame))
+	alpha := 255 * (sceneEffect.Tick / sceneEffect.Frame)
 	img := generateFadeImage(sceneEffect, w, h, alpha)
 	screen.DrawImage(img, nil)
 }
